fixed crashing and updated conditions
